+++ src/routes/content/content.ts
@@ -11,7 +11,7 @@ import type { Knex } from 'knex';
 import express from 'express';
 import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc';
-import { drop, flattenDeep, intersection, uniq } from 'es-toolkit/array';
+import { drop, intersection } from 'es-toolkit/array';
 import { omit, pick } from 'es-toolkit/object';
 import { v4 as uuid } from 'uuid';
 
@@ -27,7 +27,7 @@ import {
   handleRelationLists,
 } from '../../helpers/content/content';
 import { RequestException } from '../../helpers/error/error';
-import { readFile, removeFile } from '../../helpers/fs/fs';
+import { readFile } from '../../helpers/fs/fs';
 import { lockExpired } from '../../helpers/lock/lock';
 import { getRootUrl } from '../../helpers/url/url';
 import { mapAsync, uniqueId, bytesToNumber } from '../../helpers/utils/utils';
@@ -626,9 +626,10 @@ export default [
       };
 
       // Handle files, images and relation lists
-      json = await handleFiles(json, type, trx);
-      json = await handleImages(json, type, trx);
-      json = await handleRelationLists(json, req.type);
+      const schema = type._schema;
+      json = await handleFiles(json, schema, trx);
+      json = await handleImages(json, schema, trx);
+      json = await handleRelationLists(json, schema);
 
       // Create new document
       let document = Document.fromJson({
@@ -811,9 +812,10 @@ export default [
           omitProperties,
         ),
       };
-      json = await handleFiles(json, req.type, trx);
-      json = await handleImages(json, req.type, trx);
-      json = await handleRelationLists(json, req.type);
+      const schema = config.settings.userschema(req);
+      json = await handleFiles(json, schema, trx);
+      json = await handleImages(json, schema, trx);
+      json = await handleRelationLists(json, schema);
 
       // Create new version
       const modified = dayjs.utc().format();